# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/k8s/ingress/create": {
            "post": {
                "description": "创建ingress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingress"
                ],
                "summary": "创建ingress",
                "operationId": "/api/k8s/ingress/create",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kubeDto.IngressCreteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"新增成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/ingress/del": {
            "delete": {
                "description": "删除ingress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingress"
                ],
                "summary": "删除ingress",
                "operationId": "/api/k8s/ingress/del",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ingress名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"删除成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/ingress/detail": {
            "get": {
                "description": "获取ingress详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingress"
                ],
                "summary": "获取ingress详情",
                "operationId": "/api/k8s/ingress/detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ingress名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\":\"\"  }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/k8s/ingress/list": {
            "get": {
                "description": "查看ingress列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingress"
                ],
                "summary": "查看ingress列表",
                "operationId": "/api/k8s/ingress/list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "过滤",
                        "name": "filter_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\":\"\"  }",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
        "/api/k8s/ingress/update": {
            "put": {
                "description": "更新ingress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingress"
                ],
                "summary": "更新ingress",
                "operationId": "/api/k8s/ingress/update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ingress名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "更新内容",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, msg=\"\",\"data\": \"更新成功}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "kubeDto.HttpPath": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "path_type": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "service_port": {
                    "type": "integer"
                }
            }
        },
        "kubeDto.IngressCreteInput": {
            "type": "object",
            "properties": {
                "hosts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/kubeDto.HttpPath"
                        }
                    }
                },
                "label": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "real_err": {
                    "type": "string"
                }
            }
        },
...
    }
}